Separated DataState logic from store

+++ src/lib/app-infrastructure/collection-app/smart-store.svelte.ts
@@ -1,12 +1,8 @@
 import { browser } from '$app/environment';
 import { getDeviceId } from '$lib/engine/storage/local/client-info-repository';
-import { Value } from 'sass';
 import { deepAssign } from '../deep-assign';
-import { stampAppRecord } from './data';
-import { getErrorMessage } from '$lib/engine/general-js-ts/extract-error-message';
-import { keyBoardFocusNavigatedNode } from '$lib/engine/keyboard-navigation/navigation-utils';
 import { Cache } from './cache';
-import { error } from '@sveltejs/kit';
+import { stampAppRecord } from './data';
 
 export type SmartStoreOptions<T> = {
 	loadNotFoundBehavior: { action: 'error' } | { action: 'create-new'; createObj: () => T };
@@ -31,7 +27,9 @@ function abortable<T>(p: Promise<T>): { abort: () => void; abortablePromise: Abo
 		abort: () => (isAborted = true),
 		abortablePromise: {
 			promise: p,
-			isAborted
+			get isAborted() {
+				return isAborted;
+			}
 		}
 	};
 }
@@ -40,6 +38,7 @@ export class SmartStore<T> {
 	#context: CollectionAppContext;
 	#record: RecordI<T>;
 	#dataState: AppDataState;
+
 	#recordAdapter: CollectionAppRecordAdapter<T, SyncableAppRecordMetadata>;
 	#options: SmartStoreOptions<T>;
 	#repository: AppRecordRepo<T, SyncableAppRecordMetadata, CollectionAppError>;
@@ -47,12 +46,18 @@ export class SmartStore<T> {
 
 	#reloadAbort: (() => void) | undefined;
 
+	#saveOpCouter = 0;
+
 	get data() {
 		return this.#record.data;
 	}
 
 	get dataState() {
 		return this.#dataState;
+	}
+
+	get dataKey() {
+		return this.#record?.key;
 	}
 
 	constructor(
@@ -77,18 +82,41 @@ export class SmartStore<T> {
 	}
 
 	async save(): Promise<StoreSaveActionResult> {
+		let contextSnapshot = $state.snapshot(this.#context);
+		// TODO AZ unify logic with saveAs
+		let dataState = $state.snapshot(this.#dataState);
+		if (dataState.kind !== 'ready' && dataState.kind !== 'saving') {
+			return {
+				ok: true,
+				value: { kind: 'another-operation-in-progress', currentOperation: dataState }
+			};
+		}
+
 		stampAppRecord(getDeviceId(), this.#record.meta);
 		let saveData = $state.snapshot(this.#record) as AppRecord<T, SyncableAppRecordMetadata>;
+
+		let currentSaveOp = ++this.#saveOpCouter;
+
+		let recordSnapshot = $state.snapshot(this.#record);
 
 		this.#dataState = { kind: 'saving', key: saveData.key };
 		let res = await this.#repository.update(this.#context, saveData);
 
 		if (res.ok) {
-			let prevItemKey = this.#context.itemKey;
-			let newItemKey = this.#record.key;
+			let prevItemKey = contextSnapshot.itemKey;
+			let newItemKey = recordSnapshot.key;
 
-			this.#cache.setOrUpdateKey(newItemKey, saveData, prevItemKey);
-			this.#dataState = { kind: 'ready', key: newItemKey };
+			await this.#cache.setOrUpdateKey(newItemKey, saveData, prevItemKey);
+
+			const dataStateAfterOp = $state.snapshot(this.#dataState) as AppDataState;
+			if (
+				dataStateAfterOp.kind !== 'loading' &&
+				currentSaveOp === this.#saveOpCouter &&
+				// this is to check if context have changed since save start to prevent false reporting on context.itemKey
+				contextSnapshot.itemKey !== this.#context.itemKey
+			) {
+				this.#dataState = { kind: 'ready', key: newItemKey };
+			}
 
 			if (prevItemKey !== newItemKey) {
 				return {
@@ -109,6 +137,16 @@ export class SmartStore<T> {
 	}
 
 	async saveAs(context: CollectionAppContext, newItemKey: string): Promise<StoreSaveActionResult> {
+		let dataState = $state.snapshot(this.#dataState);
+		if (dataState.kind !== 'ready' && dataState.kind !== 'saving') {
+			return {
+				ok: true,
+				value: { kind: 'another-operation-in-progress', currentOperation: dataState }
+			};
+		}
+
+		let contextSnapshot = $state.snapshot(context);
+
 		let prevItemKey = this.#record.key;
 
 		this.#record.key = newItemKey;
@@ -117,15 +155,24 @@ export class SmartStore<T> {
 		let saveData = $state.snapshot(this.#record.data) as T;
 
 		this.#dataState = { kind: 'saving', key: this.#record.key };
+		let currentSaveOp = this.#saveOpCouter++;
 
 		console.log('Creating New Data', saveData);
 		let res = await this.#repository.create(context, saveData);
 
 		if (res.ok) {
-			this.#record = this.#recordAdapter.fromDb(res.value);
-			this.#cache.setOrUpdateKey(newItemKey, this.#record);
+			let newRecord = this.#recordAdapter.fromDb(res.value);
+			await this.#cache.setOrUpdateKey(newItemKey, newRecord);
 
-			this.#dataState = { kind: 'ready', key: newItemKey };
+			if (contextSnapshot.itemKey === this.#context.itemKey) {
+				// this check is to prevent changing of working record during context changes
+				this.#record = newRecord;
+			}
+
+			const dataStateAfterOp = $state.snapshot(this.#dataState) as AppDataState;
+			if (dataStateAfterOp.kind !== 'loading' && currentSaveOp === this.#saveOpCouter) {
+				this.#dataState = { kind: 'ready', key: newItemKey };
+			}
 
 			return { ok: true, value: { kind: 'create', newItemKey: newItemKey } };
 		} else {
@@ -137,31 +184,54 @@ export class SmartStore<T> {
 	}
 
 	async delete(context: CollectionAppContext): Promise<CollectionAppBlankResult> {
+		// TODO AZ add result here also
+
+		if (this.#dataState.kind !== 'ready') return { ok: true, value: undefined };
+
 		let res = await this.#repository.delete(context, this.#record);
 
 		if (res.ok) {
-			this.#cache.delete(this.#record.key);
+			await this.#cache.delete(this.#record.key);
 			return { ok: true, value: undefined };
 		} else {
 			return res;
 		}
 	}
 
+	// TODO AZ Return ReloadResult
 	async reload(
-		context: CollectionAppContext,
+		newContext: CollectionAppContext,
 		placeHolderValue?: T,
 		options?: SmartStoreOptions<T>
 	) {
+		let recordSnapshot = $state.snapshot(this.#record);
+		let prevContextSnapshot = $state.snapshot(this.#context);
+		let prevItemKey = recordSnapshot.key;
+		let newContextSnapshot = $state.snapshot(newContext);
+
 		this.#reloadAbort?.();
-		this.#context = context;
+
+		this.#context = newContext;
 
 		if (options) {
 			this.#options = options;
 		}
 
-		if (context.itemKey === this.#record.key) return;
+		// TODO AZ organize this mess
+		if (newContextSnapshot.itemKey === prevItemKey) {
+			console.log(
+				'Store reload canceled - current record key matches new context:',
+				newContextSnapshot,
+				'prevItemKey',
+				prevItemKey,
+				'current record',
+				$state.snapshot(this.#record)
+			);
+			return;
+		}
 
-		this.#dataState = { kind: 'loading', key: context.itemKey };
+		console.log('Store Starting Store Reload. new Context', newContextSnapshot);
+		this.#dataState = { kind: 'loading', key: newContext.itemKey };
 		if (placeHolderValue) {
 			this.#record = this.#recordAdapter.constructRecord(placeHolderValue);
 		}
@@ -172,7 +242,7 @@ export class SmartStore<T> {
 		let loadResult = await abortablePromise.promise;
 
 		if (abortablePromise.isAborted) {
-			console.log('Store - Aborted Reload', context);
+			console.log('Store - Aborted Reload', newContextSnapshot);
 			return;
 		}
